feat: main 클로버 개수 통신값 사용으로 수정
