+++ web/app/layout.tsx
@@ -1,36 +1,24 @@
-'use client';
 import './globals.css';
-import DefaultLayout from '@/components/common/DefaultLayout';
-import { usePathname } from 'next/navigation';
 import { lufga, pretendard } from '@/fonts/fonts';
-import { routeConfigs } from '@/configs/routeConfig';
+import ClientLayoutWrapper from '@/components/common/ClientLayoutWrapper'; // 새로 만든 컴포넌트 import
+import React from 'react';
+
+export const metadata = {
+  title: 'Runky'
+};
 
 export default function RootLayout({
   children
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  const pathname = usePathname();
-
-  // URL별 뒤로가기 링크
-  const config = routeConfigs[pathname];
   return (
     <html
       lang="ko"
       className={`${lufga.variable} ${pretendard.variable} antialiased`}
     >
-      <head>
-        <title>Runky</title>
-      </head>
       <body className="bg-background pt-[env(safe-area-inset-top)] pb-[env(safe-area-inset-bottom)] text-white">
-        <DefaultLayout
-          showHeader={config?.showHeader}
-          title={config?.title}
-          backHref={config?.backHref}
-          showSaveButton={config?.showSaveButton}
-        >
-          {children}
-        </DefaultLayout>
+        <ClientLayoutWrapper>{children}</ClientLayoutWrapper>
       </body>
     </html>
   );
